requirements.txt
# Core LangChain stack
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.23

# OpenAI API
openai==1.77.0

# FAISS for vector search
faiss-cpu==1.13.2

# Web & UI
streamlit==1.44.1

# Environment variables
python-dotenv==1.1.0

# PDF handling
pypdf==5.4.0

# Tokenizer compatible with langchain-openai
tiktoken==0.8.0

# Data and numeric computing
numpy>=1.24.0
pandas>=2.0.3

# Config & serialization
PyYAML>=6.0
SQLAlchemy>=2.0.0
orjson>=3.9.14

# HTTP clients
httpx>=0.28.0
aiohttp>=3.8.5

# Typing & protobuf
typing-extensions>=4.15.0
protobuf>=3.20

# Additional LangChain dependencies
langchain-core>=0.3.83,<1.0.0
langchain-text-splitters>=0.3.11,<1.0.0
langsmith>=0.3.45,<0.4
pydantic>=2.7.4,<3.0.0
requests<3,>=2